# Test perf of Request.Body.CopyToAsync()