@halter73 2016-09-20
This was originally made to test the perf of Kestrel's CopyToAsync changes. I used the default post.lua script and the /copy endpoint to test it. The RPS numbers are well withing a stddev on my laptop, but the number of gen 0 collections is far less (not that gen 0 collections are very expensive obvs).

## Before

    C:\Users\shalter\dev\benchmarks\src\Benchmarks [dev +2 ~6 -0 !]> #dev-to-body
    C:\Users\shalter\dev\benchmarks\src\Benchmarks [dev +2 ~6 -0 !]> dotnet run -c Release --scenarios copy --server.urls http://*:5000
    ASP.NET Core Benchmarks
    -----------------------
    Current directory: C:\Users\shalter\dev\benchmarks\src\Benchmarks
    Scenario configuration found in scenarios.json and/or command line args

    The following scenarios were enabled:
      Copy -> /copy

    Using server Kestrel
    Server GC is currently ENABLED
    Press 'C' to force GC or any other key to display GC stats

    Hosting environment: Production
    Content root path: C:\Users\shalter\dev\benchmarks\src\Benchmarks
    Now listening on: http://*:5000
    Application started. Press Ctrl+C to shut down.

    Allocated: 19.00 MB
    Gen 0: 1949, Gen 1: 44, Gen 2: 3

    Forcing GC... done!

    Allocated: 3.29 MB
    Gen 0: 1951, Gen 1: 46, Gen 2: 5
    Application is shutting down...

    [shalter@shalter-arch wrk]$ #dev-to-body
    [shalter@shalter-arch wrk]$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://win8:5000/copy
    Running 10s test @ http://win8:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    22.31ms   50.79ms 499.99ms   95.81%
        Req/Sec   626.83    112.44     0.88k    86.77%
      191754 requests in 10.04s, 28.16MB read
    Requests/sec:  19108.19
    Transfer/sec:      2.81MB
    [shalter@shalter-arch wrk]$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://win8:5000/copy
    Running 10s test @ http://win8:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    12.82ms   15.33ms 254.07ms   98.57%
        Req/Sec   695.38    174.38     1.40k    83.36%
      201116 requests in 10.03s, 29.54MB read
    Requests/sec:  20053.62
    Transfer/sec:      2.95MB
    [shalter@shalter-arch wrk]$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://win8:5000/copy
    Running 10s test @ http://win8:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    14.22ms   16.96ms 299.23ms   98.27%
        Req/Sec   638.15     86.92     0.85k    79.90%
      201922 requests in 10.03s, 29.66MB read
    Requests/sec:  20141.49
    Transfer/sec:      2.96MB

## After

    C:\Users\shalter\dev\benchmarks\src\Benchmarks [dev +2 ~6 -0 !]> #copyless-to-body
    C:\Users\shalter\dev\benchmarks\src\Benchmarks [dev +2 ~6 -0 !]> dotnet run -c Release --scenarios copy --server.urls http://*:5000
    ASP.NET Core Benchmarks
    -----------------------
    Current directory: C:\Users\shalter\dev\benchmarks\src\Benchmarks
    Scenario configuration found in scenarios.json and/or command line args

    The following scenarios were enabled:
      Copy -> /copy

    Using server Kestrel
    Server GC is currently ENABLED
    Press 'C' to force GC or any other key to display GC stats

    Hosting environment: Production
    Content root path: C:\Users\shalter\dev\benchmarks\src\Benchmarks
    Now listening on: http://*:5000
    Application started. Press Ctrl+C to shut down.

    Allocated: 19.01 MB
    Gen 0: 12, Gen 1: 3, Gen 2: 1

    Forcing GC... done!

    Allocated: 2.91 MB
    Gen 0: 14, Gen 1: 5, Gen 2: 3
    Application is shutting down...

    [shalter@shalter-arch wrk]$ #copyless-to-body
    [shalter@shalter-arch wrk]$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://win8:5000/copy
    Running 10s test @ http://win8:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    15.16ms   22.46ms 316.03ms   97.75%
        Req/Sec   644.73    126.47     1.08k    79.07%
      197206 requests in 10.08s, 28.96MB read
    Requests/sec:  19566.71
    Transfer/sec:      2.87MB
    [shalter@shalter-arch wrk]$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://win8:5000/copy
    Running 10s test @ http://win8:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    14.62ms   19.43ms 305.61ms   97.90%
        Req/Sec   643.31    101.32     0.94k    78.96%
      203276 requests in 10.03s, 29.85MB read
    Requests/sec:  20276.71
    Transfer/sec:      2.98MB
    [shalter@shalter-arch wrk]$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://win8:5000/copy
    Running 10s test @ http://win8:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    14.21ms   18.05ms 280.33ms   97.98%
        Req/Sec   652.34    120.25     1.33k    85.88%
      204888 requests in 10.03s, 30.09MB read
    Requests/sec:  20431.49
    Transfer/sec:      3.00MB