@halter73 2016-09-20
https://github.com/aspnet/KestrelHttpServer/pull/1105