@halter73 2016-09-20
The difference is 1949 vs 12 gen 0 collections, 44 vs 3  gen 1 collections and 3 vs 1 gen 2 collections before forcing any collections.