@halter73 2016-09-20
Seeing nearly a 4x improvement in RPS when we run this benchmark in our perflab with pipelining! 🎉 

## Before

    C:\code\github\aspnet\benchmarks\src\Benchmarks>dotnet run -c Release --scenarios copy --server.urls http://*:5000
    ASP.NET Core Benchmarks
    -----------------------
    Current directory: C:\code\github\aspnet\benchmarks\src\Benchmarks
    Scenario configuration found in scenarios.json and/or command line args

    The following scenarios were enabled:
      Copy -> /copy

    Using server Kestrel
    Server GC is currently ENABLED
    Press 'C' to force GC or any other key to display GC stats

    Hosting environment: Production
    Content root path: C:\code\github\aspnet\benchmarks\src\Benchmarks
    Now listening on: http://*:5000
    Application started. Press Ctrl+C to shut down.

    Allocated: 106.49 MB
    Gen 0: 2313, Gen 1: 137, Gen 2: 7

    Forcing GC... done!

    Allocated: 5.63 MB
    Gen 0: 2315, Gen 1: 139, Gen 2: 9
    Application is shutting down...

    adx@adx-3u-svr02:~/code/github/wg/wrk$ #dev
    adx@adx-3u-svr02:~/code/github/wg/wrk$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://10.0.0.100:5000/copy -- 16
    Running 10s test @ http://10.0.0.100:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    16.91ms   17.48ms 185.50ms   89.28%
        Req/Sec     6.49k     1.16k   11.37k    86.32%
      2003141 requests in 10.10s, 294.19MB read
      Socket errors: connect 0, read 0, write 300, timeout 0
    Requests/sec: 198313.55
    Transfer/sec:     29.13MB
    adx@adx-3u-svr02:~/code/github/wg/wrk$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://10.0.0.100:5000/copy -- 16
    Running 10s test @ http://10.0.0.100:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    16.09ms   15.91ms 249.78ms   88.76%
        Req/Sec     6.52k     1.20k   19.74k    86.37%
      2046536 requests in 10.10s, 300.57MB read
      Socket errors: connect 0, read 0, write 300, timeout 0
    Requests/sec: 202629.87
    Transfer/sec:     29.76MB
    adx@adx-3u-svr02:~/code/github/wg/wrk$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://10.0.0.100:5000/copy -- 16
    Running 10s test @ http://10.0.0.100:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    16.10ms   16.50ms 224.28ms   89.41%
        Req/Sec     6.55k     1.28k   22.52k    85.54%
      2040074 requests in 10.10s, 299.62MB read
      Socket errors: connect 0, read 0, write 300, timeout 0
    Requests/sec: 201982.69
    Transfer/sec:     29.66MB


## After

    C:\code\github\aspnet\benchmarks\src\Benchmarks>dotnet run -c Release --scenarios copy --server.urls http://*:5000
    ASP.NET Core Benchmarks
    -----------------------
    Current directory: C:\code\github\aspnet\benchmarks\src\Benchmarks
    Scenario configuration found in scenarios.json and/or command line args

    The following scenarios were enabled:
      Copy -> /copy

    Using server Kestrel
    Server GC is currently ENABLED
    Press 'C' to force GC or any other key to display GC stats

    Hosting environment: Production
    Content root path: C:\code\github\aspnet\benchmarks\src\Benchmarks
    Now listening on: http://*:5000
    Application started. Press Ctrl+C to shut down.

    Allocated: 79.96 MB
    Gen 0: 90, Gen 1: 19, Gen 2: 2

    Forcing GC... done!

    Allocated: 5.59 MB
    Gen 0: 92, Gen 1: 21, Gen 2: 4
    Application is shutting down...

    adx@adx-3u-svr02:~/code/github/wg/wrk$ #copyless
    adx@adx-3u-svr02:~/code/github/wg/wrk$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://10.0.0.100:5000/copy -- 16
    Running 10s test @ http://10.0.0.100:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    14.69ms   32.19ms 293.26ms   89.46%
        Req/Sec    25.18k     4.13k   44.36k    79.88%
      7780714 requests in 10.10s, 1.12GB read
      Socket errors: connect 0, read 0, write 300, timeout 0
    Requests/sec: 770481.24
    Transfer/sec:    113.16MB
    adx@adx-3u-svr02:~/code/github/wg/wrk$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://10.0.0.100:5000/copy -- 16
    Running 10s test @ http://10.0.0.100:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    13.84ms   32.50ms 350.60ms   89.70%
        Req/Sec    25.27k     3.69k   47.68k    75.79%
      7999287 requests in 10.10s, 1.15GB read
      Socket errors: connect 0, read 0, write 300, timeout 0
    Requests/sec: 791996.46
    Transfer/sec:    116.32MB
    adx@adx-3u-svr02:~/code/github/wg/wrk$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://10.0.0.100:5000/copy -- 16
    Running 10s test @ http://10.0.0.100:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency    12.35ms   30.56ms 388.73ms   91.04%
        Req/Sec    24.98k     3.47k   74.09k    81.38%
      7916825 requests in 10.10s, 1.14GB read
      Socket errors: connect 0, read 0, write 600, timeout 0
    Requests/sec: 783893.03
    Transfer/sec:    115.13MB

I'm going to add the pipelined post.lua I used to the PR.