@halter73 2016-09-20
About a 2x improvement without pipelining:


## Before
    adx@adx-3u-svr02:~/code/github/wg/wrk$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://10.0.0.100:5000/copy -- 1
    Running 10s test @ http://10.0.0.100:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     2.52ms    7.27ms 141.25ms   97.65%
        Req/Sec     5.01k     0.85k    8.83k    90.98%
      1550953 requests in 10.10s, 227.78MB read
      Socket errors: connect 0, read 0, write 489, timeout 0
    Requests/sec: 153564.10
    Transfer/sec:     22.55MB
    adx@adx-3u-svr02:~/code/github/wg/wrk$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://10.0.0.100:5000/copy -- 1
    Running 10s test @ http://10.0.0.100:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     1.90ms    1.98ms  70.69ms   91.72%
        Req/Sec     5.02k   437.31     8.73k    83.21%
      1592348 requests in 10.10s, 233.86MB read
      Socket errors: connect 0, read 0, write 300, timeout 0
    Requests/sec: 157673.33
    Transfer/sec:     23.16MB


## After

    adx@adx-3u-svr02:~/code/github/wg/wrk$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://10.0.0.100:5000/copy -- 1
    Running 10s test @ http://10.0.0.100:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     1.85ms    8.30ms 142.05ms   97.64%
        Req/Sec    10.38k     1.51k   17.12k    91.58%
      3219496 requests in 10.10s, 472.83MB read
      Socket errors: connect 0, read 0, write 536, timeout 0
    Requests/sec: 318763.24
    Transfer/sec:     46.82MB
    adx@adx-3u-svr02:~/code/github/wg/wrk$ ./wrk -c 256 -t 32 -d 10 -s post.lua http://10.0.0.100:5000/copy -- 1
    Running 10s test @ http://10.0.0.100:5000/copy
      32 threads and 256 connections
      Thread Stats   Avg      Stdev     Max   +/- Stdev
        Latency     2.60ms    9.18ms 160.04ms   95.40%
        Req/Sec    10.22k   839.38    13.82k    65.95%
      3285931 requests in 10.10s, 482.59MB read
    Requests/sec: 325363.86
    Transfer/sec:     47.78MB